move more stuff to cudf.api.types
