+++ python/cudf/cudf/_lib/parquet.pyx
@@ -11,7 +11,8 @@ import json
 from cython.operator import dereference
 import numpy as np
 
-from cudf.utils.dtypes import np_to_pa_dtype, is_categorical_dtype
+from cudf.utils.dtypes import np_to_pa_dtype
+from cudf.api.types import is_categorical_dtype
 from libc.stdlib cimport free
 from libc.stdint cimport uint8_t
 from libcpp.memory cimport shared_ptr, unique_ptr, make_unique
