+++ python/cudf/cudf/core/column/lists.py
@@ -5,7 +5,7 @@ import cudf
 from cudf.core.column import ColumnBase
 from cudf.core.column.methods import ColumnMethodsMixin
 from cudf.core.dtypes import ListDtype
-from cudf.utils.dtypes import is_list_dtype
+from cudf.api.types import is_list_dtype
 from cudf.utils.utils import buffers_from_pyarrow
