+++ python/cudf/cudf/core/series.py
@@ -42,14 +42,13 @@ from cudf.utils import cudautils, ioutils, utils
 from cudf.utils.docutils import copy_docstring
 from cudf.utils.dtypes import (
     can_convert_to_column,
-    is_list_dtype,
     is_list_like,
     is_mixed_with_object_dtype,
     is_scalar,
-    is_string_dtype,
     min_scalar_type,
     numeric_normalize_types,
 )
+from cudf.api.types import is_list_dtype, is_string_dtype
 
 
 class Series(Frame, Serializable):
